renamed and cleaned up code
